refactor(follow): move follow logic to own controller and service
