+++ src/main/java/ispp/project/dondesiempre/modules/stores/controllers/StoreController.java
@@ -1,8 +1,5 @@
 package ispp.project.dondesiempre.modules.stores.controllers;
 
-import ispp.project.dondesiempre.modules.auth.services.UserService;
-import ispp.project.dondesiempre.modules.clients.models.Client;
-import ispp.project.dondesiempre.modules.follows.dtos.StoreFollowerDTO;
 import ispp.project.dondesiempre.modules.stores.dtos.StoreDTO;
 import ispp.project.dondesiempre.modules.stores.dtos.StoreUpdateDTO;
 import ispp.project.dondesiempre.modules.stores.services.StoreService;
@@ -11,10 +8,8 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -27,7 +22,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class StoreController {
   private final StoreService storeService;
-  private final UserService userService;
 
   @GetMapping("/stores")
   @ResponseStatus(HttpStatus.OK)
@@ -60,35 +54,5 @@ public class StoreController {
   public ResponseEntity<StoreDTO> updateStore(
       @PathVariable("id") UUID id, @RequestBody StoreUpdateDTO dto) {
     return new ResponseEntity<>(storeService.updateStore(id, dto), HttpStatus.OK);
-  }
-
-  @PostMapping("/stores/{storeId}/followers")
-  @ResponseStatus(HttpStatus.CREATED)
-  public ResponseEntity<String> followStore(@PathVariable("storeId") UUID storeId) {
-    storeService.followStore(storeId);
-    return new ResponseEntity<>("Store followed successfully", HttpStatus.CREATED);
-  }
-
-  @DeleteMapping("/stores/{storeId}/followers/me")
-  @ResponseStatus(HttpStatus.OK)
-  public ResponseEntity<String> unfollowStore(@PathVariable("storeId") UUID storeId) {
-    storeService.unfollowStore(storeId);
-    return new ResponseEntity<>("Store unfollowed successfully.", HttpStatus.OK);
-  }
-
-  @GetMapping("/clients/me/followed-stores")
-  @ResponseStatus(HttpStatus.OK)
-  public ResponseEntity<List<StoreDTO>> getMyFollowedStores() {
-    List<StoreDTO> followedStores =
-        storeService.getMyFollowedStores().stream().map(storeService::toDTO).toList();
-    return new ResponseEntity<>(followedStores, HttpStatus.OK);
-  }
-
-  @GetMapping("/stores/{storeId}/followers/me")
-  public ResponseEntity<StoreFollowerDTO> checkIfIFollowStore(@PathVariable UUID storeId) {
-    Client currentClient = userService.getCurrentClient();
-    boolean follows = storeService.checkIfClientFollowsStore(currentClient.getId(), storeId);
-    return new ResponseEntity<>(
-        new StoreFollowerDTO(currentClient.getId(), storeId, follows), HttpStatus.OK);
   }
 }
