+++ src/test/java/ispp/project/dondesiempre/modules/stores/services/StoreServiceTest.java
@@ -2,23 +2,18 @@ package ispp.project.dondesiempre.modules.stores.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyDouble;
 import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import ispp.project.dondesiempre.mockEntities.StoreMockEntities;
 import ispp.project.dondesiempre.modules.auth.services.AuthService;
-import ispp.project.dondesiempre.modules.auth.services.UserService;
 import ispp.project.dondesiempre.modules.clients.models.Client;
 import ispp.project.dondesiempre.modules.common.exceptions.InvalidBoundingBoxException;
 import ispp.project.dondesiempre.modules.common.exceptions.ResourceNotFoundException;
 import ispp.project.dondesiempre.modules.follows.models.StoreFollower;
-import ispp.project.dondesiempre.modules.follows.repositories.StoreFollowerRepository;
 import ispp.project.dondesiempre.modules.stores.dtos.StoreDTO;
 import ispp.project.dondesiempre.modules.stores.models.SocialNetwork;
 import ispp.project.dondesiempre.modules.stores.models.Store;
@@ -42,8 +37,6 @@ public class StoreServiceTest {
 
   @Mock private StoreRepository storeRepository;
   @Mock private StoreSocialNetworkRepository socialNetworkRepository;
-  @Mock private StoreFollowerRepository storeFollowerRepository;
-  @Mock private UserService userService;
   @Mock private AuthService authService;
 
   @Mock private ApplicationContext applicationContext;
@@ -138,100 +131,5 @@ public class StoreServiceTest {
     assertEquals(1, result.getSocialNetworks().size());
     assertEquals("instagram", result.getSocialNetworks().get(0).getName());
     assertEquals("https://instagram.com/store", result.getSocialNetworks().get(0).getLink());
-  }
-
-  @Test
-  void followStore_shouldFollowStore_whenStoreExists() {
-    when(userService.getCurrentClient()).thenReturn(TEST_CLIENT);
-    when(storeRepository.findById(TEST_STORE_ID)).thenReturn(Optional.of(TEST_STORE));
-    when(storeFollowerRepository.save(any())).thenReturn(TEST_FOLLOWER);
-    when(applicationContext.getBean(StoreService.class)).thenReturn(storeService);
-
-    StoreFollower follow = storeService.followStore(TEST_STORE_ID);
-
-    assertEquals(follow.getClient().getId(), TEST_FOLLOWER.getClient().getId());
-    assertEquals(follow.getStore().getId(), TEST_FOLLOWER.getStore().getId());
-    verify(storeFollowerRepository).save(follow);
-    verify(applicationContext).getBean(StoreService.class);
-  }
-
-  @Test
-  void unfollowStore_shouldThrowNotFound_whenNotPreviouslyFollowed() {
-    when(userService.getCurrentClient()).thenReturn(TEST_CLIENT);
-    when(storeFollowerRepository.findByClientIdAndStoreId(TEST_CLIENT.getId(), TEST_STORE_ID))
-        .thenReturn(Optional.empty());
-
-    assertThrows(ResourceNotFoundException.class, () -> storeService.unfollowStore(TEST_STORE_ID));
-    verify(storeFollowerRepository, never()).delete(any());
-  }
-
-  @Test
-  void unfollowStore_shouldUnfollowStore_whenPreviouslyFollowed() {
-    when(userService.getCurrentClient()).thenReturn(TEST_CLIENT);
-    when(storeFollowerRepository.findByClientIdAndStoreId(TEST_CLIENT.getId(), TEST_STORE_ID))
-        .thenReturn(Optional.of(TEST_FOLLOWER));
-    doNothing().when(storeFollowerRepository).delete(any());
-
-    storeService.unfollowStore(TEST_STORE_ID);
-
-    verify(storeFollowerRepository, times(1)).delete(TEST_FOLLOWER);
-  }
-
-  @Test
-  void getMyFollowedStores_shouldGetZeroStores_whenNotFollowingAnyStore() {
-    when(userService.getCurrentClient()).thenReturn(TEST_CLIENT);
-    when(storeFollowerRepository.findByClientId(TEST_CLIENT.getId())).thenReturn(List.of());
-
-    List<Store> stores = storeService.getMyFollowedStores();
-
-    assertEquals(stores.size(), 0);
-  }
-
-  @Test
-  void getMyFollowedStores_shouldGetOneStore_whenFollowingOneStore() {
-    when(userService.getCurrentClient()).thenReturn(TEST_CLIENT);
-    when(storeFollowerRepository.findByClientId(TEST_CLIENT.getId()))
-        .thenReturn(List.of(TEST_FOLLOWER));
-
-    List<Store> stores = storeService.getMyFollowedStores();
-
-    assertEquals(stores.size(), 1);
-    assertEquals(stores.get(0).getId(), new StoreDTO(TEST_STORE).getId());
-  }
-
-  @Test
-  void getMyFollowedStores_shouldGetMultipleStores_whenFollowingMultipleStores() {
-    when(userService.getCurrentClient()).thenReturn(TEST_CLIENT);
-    when(storeFollowerRepository.findByClientId(TEST_CLIENT.getId()))
-        .thenReturn(List.of(TEST_FOLLOWER, TEST_FOLLOWER, TEST_FOLLOWER));
-
-    List<Store> stores = storeService.getMyFollowedStores();
-
-    assertEquals(stores.size(), 3);
-    assertEquals(stores.get(0).getId(), TEST_STORE.getId());
-  }
-
-  @Test
-  void checkIfIFollowStore_shouldReturnTrue_whenFollowingStore() {
-    when(storeFollowerRepository.existsByClientIdAndStoreId(TEST_CLIENT.getId(), TEST_STORE_ID))
-        .thenReturn(true);
-
-    boolean res = storeService.checkIfClientFollowsStore(TEST_CLIENT.getId(), TEST_STORE_ID);
-
-    verify(storeFollowerRepository, times(1))
-        .existsByClientIdAndStoreId(TEST_CLIENT.getId(), TEST_STORE_ID);
-    assertEquals(res, true);
-  }
-
-  @Test
-  void checkIfIFollowStore_shouldReturnFalse_whenNotFollowingStore() {
-    when(storeFollowerRepository.existsByClientIdAndStoreId(TEST_CLIENT.getId(), TEST_STORE_ID))
-        .thenReturn(false);
-
-    boolean res = storeService.checkIfClientFollowsStore(TEST_CLIENT.getId(), TEST_STORE_ID);
-
-    verify(storeFollowerRepository, times(1))
-        .existsByClientIdAndStoreId(TEST_CLIENT.getId(), TEST_STORE_ID);
-    assertEquals(res, false);
   }
 }
